+++ libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: mbonati <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 14:00:21 by mbonati           #+#    #+#             */
-/*   Updated: 2018/11/19 14:48:38 by mbonati          ###   ########.fr       */
+/*   Updated: 2018/11/22 22:54:43 by mbonati          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] && s2[i] && s1[i] == s2[i])
