+++ libft/ft_strequ.c
@@ -6,7 +6,7 @@
 /*   By: mbonati <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 17:49:28 by mbonati           #+#    #+#             */
-/*   Updated: 2018/11/16 20:18:01 by mbonati          ###   ########.fr       */
+/*   Updated: 2018/11/22 22:55:28 by mbonati          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strequ(char const *s1, char const *s2)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s1[i] && s2[i])
