+++ libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: mbonati <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/13 22:28:26 by mbonati           #+#    #+#             */
-/*   Updated: 2018/11/17 22:41:21 by mbonati          ###   ########.fr       */
+/*   Updated: 2018/11/22 22:56:11 by mbonati          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
